requirements.txt
openai
python-dotenv
pyperclip
